long long -> int64_t, unsigned long long -> uint64_t

+++ src/jsoncons/json_input_handler.hpp
@@ -258,11 +258,11 @@ private:
     {
     }
 
-    void do_integer_value(long long, const basic_parsing_context<CharT>&) override
+    void do_integer_value(int64_t, const basic_parsing_context<CharT>&) override
     {
     }
 
-    void do_uinteger_value(unsigned long long, const basic_parsing_context<CharT>&) override
+    void do_uinteger_value(uint64_t, const basic_parsing_context<CharT>&) override
     {
     }
